Update point tally functions
